refactor: split alt-view-interfaces, rename registry key to customOrderedView
- Extract AlternateViewProps/AlternateViewEditorProps to config/alt-view-interfaces.ts
- Registry only contains component mapping
- Rename view ID from battingOrder to customOrderedView (label stays 'Batting Order')
