+++ config/alternate-view-registry.ts
@@ -1,29 +1,7 @@
 import type { ComponentType } from "react";
 import CustomOrderedView from "@/components/sports/session-alt-views/custom-ordered-view";
 import CustomOrderedEditor from "@/components/sports/session-alt-views/custom-ordered-editor";
-import type { SignupRow } from "@/components/sports/session-signups-table";
-
-/** Props passed to every alternate view component. */
-export interface AlternateViewProps {
-    signups: SignupRow[];
-    teamMemberIds: Set<string>;
-    playerCap: number | null;
-    currentUserId?: string | null;
-    /** The stored data for this view (from alt_session_views[viewId]). */
-    viewData: unknown;
-}
-
-/** Props passed to every alternate view editor component. */
-export interface AlternateViewEditorProps {
-    sport: string;
-    sessionId: string;
-    viewId: string;
-    signups: SignupRow[];
-    teamMemberIds: Set<string>;
-    /** Current saved data for this view. */
-    viewData: unknown;
-    onSaved: () => void;
-}
+import type { AlternateViewProps, AlternateViewEditorProps } from "./alt-view-interfaces";
 
 interface AlternateViewEntry {
     ViewComponent: ComponentType<AlternateViewProps>;
@@ -35,7 +13,7 @@ interface AlternateViewEntry {
  * Same pattern as admin-tab-registry — generic code never imports sport-specific folders.
  */
 const alternateViewRegistry: Record<string, AlternateViewEntry> = {
-    battingOrder: {
+    customOrderedView: {
         ViewComponent: CustomOrderedView,
         EditorComponent: CustomOrderedEditor,
     },
